Extract moveElements into extensions folder
